backend/send-rsvp/requirements.txt
pydantic[email]==2.5.0
psycopg2-binary==2.9.9